requirements.txt
fastapi[all]
uvicorn[standard]
langchain
langchain-core
langchain-community
langchain-text-splitters
langchain-google-genai
chromadb
pydantic
pydantic-settings
python-dotenv
